fix: prevent crash on malformed URI
I was migrating from a create-react-app project, and my application wouldn't start. Looking at the terminal, I noticed vite had crashed, and it wasn't obvious why. Diving in, it looks like `decodeURI(...)` throws an error on bad URLs (which makes sense), but we shouldn't crash the server when that happens.

This PR adds a `try/catch` around the `decodeURI` call. It also gives a specific message when it encounters `%PUBLIC.URL%`, pointing folks to the vite docs for how it handles `index.html`