docs: add note about the .nojekyll file for static site deployment
I wanted to keep the text short but I noticed the following as well:

Even if I picked the `docs/` subfolder as the "Source" in the repository's Github Pages settings, the .nojekyll file need to be at the root.

This isn't made very clear in Github docs, they say to add `.nojekyll` "in the root of your publishing source", so logically I would think that is the "Source" I selected (docs/).

Context: I have a small Vite app with a `gh-pages` branch to deploy. I simply added a new `docs/` folder into that branch only, and copy the contents of dist/ into it before pushing the branch.

So if I did this, it didn't work;:

```
index.html
docs/
   .nojekyll
   index.html
```

Somehow, Github Pages published the index.html at the root (the one that poins to `/src/main.ts`), even though I configured it to publish from docs/.

This worked:

```
.nojekyll
index.html
docs/
   index.html
```

So TLDR I guess what they mean is the file should always be at the root of the repo, regardless what "Source" you pick.