fix: CSS dependencies are tracked incorrectly when base is set
The change in https://github.com/vitejs/vite/pull/4267 caused all CSS
dependencies of files that are not CSS (Tailwind CSS JIT deps) to be
tracked with the `base` prefix in the module graph.

URLs in the module graph are normalized and are never prefixed with base
and as a result changes to those files were not causing the CSS module
to be invalidated.

See https://github.com/tailwindlabs/tailwindcss/issues/4978#issuecomment-897620776