fix(dev): handle windows relative path handling
On Windows, when a folder is on a virtual drive (e.g. `subst X: C:/dev`),
then symlink resolving will end up causing the root of the project
and the imported files to be on separate drives.

This leads into a situations where `path.relative(root, entryFile)` does
not return a relative path.

Updates #4635