fix(plugin-legacy): Restore guessable polyfill id
The name of the polyfill file is determined by a hardcoded string.
Some packages based on this plugin rely on this predefined name.
This name was changed in 3127219bfa4129538fe5448c86b7e9aa1fe4d9ea
to prevent resolving files which don't exist. This commit changes
that behaviour to restore the predefined name.