fix(css plugin): fix the problem caused by calling vite-url-rewrite after user-defined plugin
The `vite-url-rewrite`plugin uses the input.file property to replace the url and put it in the last
execution of postcssPlugins. The problem now is that some postcss plugins that users are currently
using may make the input.file property disappear, resulting in vite not being able to handle the url
correctly and packaging the result successfully. For example
https://github.com/saionjisekai/px2rem-postcss/blob/1edc6763ad2fe8288ac6a8bda482319f006d8bb6/lib/index.js#L15