fix: Correctly process urls when they are rewritten to contain space
If a CSS file contains

background: url(folder/foo.png)

and the processing of the URL causes "folder" to be rewritten to e.g. "other folder", then the end result cannot be

background: url(other folder/foo.png)

because using space without wrapping the url with ' or " is not allowed.

This rewrites the url in this case to

background: url('other folder/foo.png')