feat(client): support "full-reload" event with any page path
In other words, stop requiring a .html extension for a selective page reload, because
that prevents certain pages from being selectively reloaded.

For example, if I send a "full-reload" event with a `path` of "/foo.html", it will
not reload clients whose location is "/foo". Similarly, sending a full-reload event
with a path of "/foo/index.html" will not reload "/foo" clients either. In this case,
only clients at "/foo/" will be reloaded.

Therefore, we should allow sending a full-reload event with a path of "/foo" to
specifically target clients at "/foo".