Use .mjs extension for library ESM bundle
As explained in #7609, when importing an ES module in Node.js, either its package needs to have `"type": "module"` set in its `package.json`, or its file extension needs to be `.mjs`. Otherwise the file will be interpreted as CommonJS and the import will fail.

By changing the extension of the ESM bundle to `.mjs`, the library bundle can be successfully used both in Node.js and in frontend bundlers.