Update default target from `safari13` to `safari14`.
From https://github.com/vitejs/vite/issues/9062, it sounds like the goal is to support ES2020 by default. However, the browser list was set to `safari13`, which means that big ints (an ES2020 feature) produced an error by default.

This PR updates from `safari13` to `safari14` (released September 2020) to support that.

Thanks for any consideration of this change!