Avoid errors when loading the overlay code in workers.
`HTMLElement` and `customElements` are not defined in e.g. web workers. However, Vite sometimes tries to inject the overlay into such workers. Ideally, it would not do so (and would e.g. find another way to notify the developer about issues). But this PR offers a change that allows such workers to run, by avoiding crashes in `overlay.ts` when it does.