Doc: Slight rename of MPA → Multiple pages
Slight rename, to for clarity. Since the section is talking about building/bundling multiple pages, not about either the routing or rendering paradigm (associated with MPA), which may in theory be separate. (This page came up as a bit confusing, when discussing various routing and rendering modes in vite-plugin-ssr.)
The doc page moveover doesn't talk about either MPA or CSR/SPA apart from this single usage, so it's better to simply not go into that conceptual model.