Delete dependent pre built virtual modules
For the old version of "esbuild", in order to flatten the dependent pre built products, the agent module will be used to process the "require ('react /jsx runtime. js')". The reason why the path was not directly modified to redirect to the entry is that the products will have a nested structure. The subsequent wrapper is also added to prevent secondary packaging.
However, for the current version of "esbuild", even if it is directly redirected to the entry module, the product name is still specified by "entryPoints" and will not result in secondary packaging or nested products. Therefore, we do not need to use the proxy module now.
After deleting and adding the above code, “vite” still runs well, and the pre built product and product size will not change.