fix: use relative paths in `sources` of transformed source maps
This refines the fix from https://github.com/vitejs/vite/pull/4985 to
turn absolute paths into relative paths for the `sources` array in
source maps for transformer outputs (and thereby completely avoids the
Windows drive letter problem). In order to minimize unintended negative
side effects, we perform this step only when the source file name is
absolute.

This addresses the issue that source files show up with an absolute path
prefix in case of Vue[^1].

Bug: https://crbug.com/1411596
Ref: https://github.com/vitejs/vite/issues/4964
Ref: https://github.com/vitejs/vite/issues/4912

[^1]: https://goo.gle/devtools-vite-interoperability