style(create-vite): use quotes for attributes consistently
As part of my changes in #12374, I unintentionally removed the quotes
for the Vite logos from the vanilla templates, perhaps because the rest
of the templates didn't need them.

While HTML5 allows no quotes for attribute values if the value doesn't contain
a space, keeping a consistent style is important for cognitive load when
reading and maintaining. Hence, I've marked this as a style change.

While not a functional change in these templates, it could be if someone
later changed the import path to contain a space (perhaps after
generating a project using the template).