fix: update svelte-check to v3.x in svelte-ts template
The svelte-ts template installs Typescript in v5.x, although svelte-preprocess, which is defined as v4.x and is a dependency of svelte-check, does not mention Typescript 5 as a valid peer dependency. Bumping svelte-check to v3 resolves the issue, because then svelte-preprocess v5 is used under the hood, which defines Typescript 5 as a valid peer dependency. Package managers, which throw an error for unmet peer dependencies, will not fail anymore.