docs: add GitHub Actions workflow example for pnpm with Vite and TypeScript
This commit provides a clear example of how to set up a GitHub Actions workflow for deploying a Vite project with TypeScript using pnpm which shows the yml file on click. 
https://gist.github.com/pierrejoubert73/902cc94d79424356a8d20be2b382e1ab

This will be a helpful resource for developers who are using these tools and struggling to get their static github page up and running.