fix: revert to overload defineConfig
Based on #13792 but with slightly more verbose function style
configuration types.
We can make the generic version pass the tests with some hack, but it
is too permissive that it would allow users to pass in unknown fields.
The overload version is much more direct and simpler.

Thanks @xiaoxiangmoe for helping me debug this.