Make react refresh work on arbitrary host name
Previously, the react-refresh script was importing the module from a hard-coded 'localhost' URL. This caused issues when the Vite server was started with the '--host' option to expose t development server to the local network for mobile debugging. In such cases, the script would fail to load on mobile devices, as they were not hosting the script on 'localhost'.

This commit refactors the react-refresh script to dynamically determine the host based on `window.location.hostname`. This allows the script to correctly load the module from the appropriate host, whether it's 'localhost' or a local network IP, resolving the issue with mobile debugging.