fix(load-fallback): only read file if request has query
I noticed when using vite-plugin-inspect that vite:load-fallback appeared to be taking up a lot of time which was odd considering how little this plugin does.
Ultimately, this is a false positive since vite-plugin-inspect can't actually measure how much cpu time is spent in a plugin for async plugins and instead it
includes arbitrary wait time waiting for the promise to resolve while other tasks are being processed.

That being said, it doesn't really make sense for the 'vite:load-fallback' plugin to call readFile for all requests as it over-inflates the time spent in this
plugin. Skipping this readFile doesn't really provide any performance benefits as rollup just does a readFile of their own if the load plugins return undefined.