feat(css): bundle css as part of optimizeDeps
This feature is intended to support component libraries that ship tree-
shakable modules that include css imports.

When building with Vite, if the js files within your dependencies,
contain css, then at build time, the css will get pulled into the app's
css bundles and will be tree-shaken as part of the build (if
sideEffects: false in the dependency and the module which imported the
css is also dropped). For dev builds, treeshaking
doesn't happen, and instead, all of the css will wind up getting
imported. The current behavior with optimizeDeps is that the css imports
are externalized and preserved in the resulting .js deps bundles. This
can result in a lot of css request being made on page load to the dev
server. While css imports are faster to process than js imports, it
still isn't ideal.

This feature extends the default preoptimized dependency bundling to
support css imports from within the dependency's .js files. Esbuild
already supports this feature, but it does so by bundling the css to a
single .css bundle that sits alongside the js bundle, and it removes all
imports to the css files from the resulting js bundle. The implied use-
case is that consumers of a library built with esbuild would need to use
a style tag, or import the css file directly.

Since we can't really change the semantics of how the library is
consumed as a result of optimizeDeps, we need to inject a single import
statement into the .js bundle if it has a corresponding cssBundle file
in the esbuild metadata file. We are just appending this import
statement to the bottom of the file so that the sourcemaps are
still functional.

The end result is that the css for these dependencies will now be
bundled into a single .css request for each dep / chunk that contains
imports to a css file.