Update index.md
We experienced a leak of some secret environment variables after someone copied this code without understanding it thoroughly. After pulling in all environment variables mistakenly, they then exposed them in the bundle by doing,

```
...
    define: {
      'process.env': env,
    },
```

While it is nice to be able to show how all of the parameters in loadEnv can be used, when there is just one code sample on the page, I think it may protect users from mistakenly loading more environment variables than they need if we don't set the third parameter to empty string here. 

The proposed change removes the third parameter but leaves the comment that describes how it can be used if needed.