fix: Allow enabling esbuild.minifyWhitespace for es format in lib mode (fix #6555)
To produce es bundles that can be tree-shaken, lib mode disables
minifyWhitespace by default when using the es format. But there are
cases where we want an es bundle that is as small as possible and that
does not need to be tree-shakable. This change allows minifyWhitespace
to be re-enabled if the user so desires.