Add `wasm` to the compressible assets regex
Wasm files are often highly compressible and it would be nice to show their compressed size in Vite builds. 

I found an earlier PR (https://github.com/vitejs/vite/pull/12485) that made this change for some other asset types, and a [comment](https://github.com/vitejs/vite/pull/12485#issuecomment-1763361804) that sounded open to the addition of wasm to the list, which I've added.