Ensure buffered error visible on all tabs.
To support showing an error on connection, we buffered up an error to send over
the HMR socket when a client connected. Previously, we would clear out this
error at that time and set it again later when we sent the payload over the
socket to the client. The problem is that multiple clients connected to HMR
would race to see that error (for example, multiple browser tabs).

This commit makes it so all tabs will see the error by only clearing out the
buffered error when any non-error update is sent over the socket. This also has
the benefit of clearing out the buffered error if an error is introduced and
fixed before the client reconnects, which can happen with even a single tab.