fix(config): workaround failure creating temporary file (take 2)
If node_modules is not writable, loadConfigFromBundledFile fails to
write its temp file and vite aborts with an exception.

This was previously fixed in #13269, but was reverted after several
reports of failure to load config with relative imports (see #13631
and #13730).

Instead of replacing the current code, try to load from string only
as a workaround, when the directory doesn't exist.

This should solve the case of unwritable node_modules when relative
imports are not used without breaking anything that works.